# app/docs/swagger.yaml
openapi: 3.1.0
info:
  title: IPST File Manager API
  version: 1.0.0
  description: API documentation for IPST File Manager
servers:
  - url: http://localhost:80
paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Registered successfully
        "400":
          description: Bad request
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Logged in successfully!
        "400":
          description: Invalid username or password
  /logout:
    get:
      summary: Logout user
      responses:
        "200":
          description: Logged out successfully!
  /folderContents:
    get:
      summary: Get folder contents
      parameters:
        - in: query
          name: path
          schema:
            type: string
          description: Path to the directory
      responses:
        "200":
          description: Directory contents
        "400":
          description: Error fetching directory contents
  /createFolder:
    post:
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Folder created successfully
        "400":
          description: Error creating folder
  /deleteFolder:
    delete:
      summary: Delete a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
      responses:
        "200":
          description: Folder deleted successfully
        "400":
          description: Error deleting folder
  /createFile:
    post:
      summary: Upload and create a new file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: File created successfully
        "400":
          description: Error creating file
  /deleteFile:
    delete:
      summary: Delete a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: File deleted successfully
        "400":
          description: Error deleting file